requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
streamlit-authenticator>=0.3.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
